Remove call to setVal() when working with reference to expression
In DeviceBlock::resolveSubcircuitInstanceParamStrings() a
reference to a parameter's expression is used to simplify
the expression.  Then it was set as the parameter's expression
with the setVal() method.  Since the function was working with
a reference to the underlying expression, the effect of calling
setVal() was to have the parameter class delete the expression
before trying to add it back as the the parameter's expresison.
On some OS's this is allowed a paramter with an expression value
to hold an expression with a null pointer.

Since resolveSubcircuitInstanceParamStrings() is working directly
on a reference to the parameter's expression, there is not a
need to call setVal() except when the expression can be resolved
to just a double.

https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/55