Fix memory leaks from parent nodes in expr AST
This is related to my work on these two issues:

"Refactor children and parent nodes in expression library AST"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/598

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24
"Treat "local" and "global" variation for UQ methods in manner similar
to Hspice"

This is similar to my code changes from yesterday, which also fixed some
memory leaks.  The issue in this commit is basically the same kind of
issue, where an RCP gets confused when it has a circular dependence.
The main difference is in the specific code involved.

Yesterday's changes had to do with the RCP version of the
"*this" pointer.  Today's changes have to do with the std::vectors of
parent nodes that I created to support issue 24.  I wanted to be able to
substitute nodes in the AST to support local parameter variation, and I
needed to track the parent nodes to do this.  Previously I only tracked
the children.

Unfortunately, I had the AST nodes themselves own std::vectors to their
parent nodes, and this created a circular ownership pattern, since
the parents already owned RCPs to the children, I couldn't have the
children also own RCPs to the parents.  With them owning each other,
the ref count process got confused, and there was no way to fix this
with custom destructors, or changing the order of destructor calls.

So, today's fix involves having the parent data structures sit outside
of the AST.  They are now owned by the newExpression class, and just
passed into AST functions as needed as local variables.  This fixes the
remaining memory leak errors in the AST.