Fix the hangs on M1 mac for the issue455 test
https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/57
"Test "issue455" hangs on our M1-based Mac Mini"

This commit hopefully fixes that M1 mac hangs.  I don't have an M1 mac
(yet), so it is a bit speculative.

The hangs were happening in the
Xyce::Device::SourceData::extractSourceFields function.   An infinite
loop was occuring in the PWL sources section.  The logical flow was
*supposed* to be (1) find metadata for the given source function and (2)
loop over that metadata and do stuff.

If (1) failed, and the metadata wasn't found, Xyce was supposed to
invoke the error manager and stop.   But in the M1 optimized build, it
was not doing that right away. It would continue for a bit, and
unfortunately, it would then try to loop over the (non-existant)
metadata.  This would happen because it would attempt to use an
invalid reference to get the size of the metadata vector.

THe solution is to make the logic more robust, and not rely on the error
handler being called promptly.

Assuming my theory is correct this commit should fix the issue.