Fixes gmin stepping when used in the spice strategy
In the spice strategy a Newton solve is performed first to compute
the DCOP solution.  If that solve fails, it proceeds to gmin stepping
to compute the DCOP solution.  The gmin stepping should be performed
in the same way as if it was the first DCOP solve.  It turns out that
for several devices the initial junction voltages were not recomputed,
so the gmin stepping was not initialized properly.  This can be traced
back to the nonlinear solver setting a firstSolveComplete_ flag after
the Newton solve, regardless of if that solve succeeded.  A simple one
line change before calling gmin stepping in the spice strategy to
reset the firstSolveComplete_ to false correctly initializes devices
that need to compute initial junction voltages.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/145