Fix for down sampling method in epressions
This is related to:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/564

I had thought this issue was finished, but it was pointed out to me that
in the regression tests I created for the issue, that two of the tests
were producing identical answers, whey they probably should have been
slightly different.

It turned out that in these 2 tests, the capability was not being
invoked.  The relevant expressions were parsed correctly, but I made an
"order-of-operations" mistake.

The down sampling function was applied in one of the tableOp
constructor functions.  However, the constructor was too early
to apply it, as the two arguments that were relevant to this
features (number of samples, and if a log scale was used)
were not resolved at this stage.

numSamples and logScaling *were* resolved at this stage if
the arguments were raw numbers.  However they were not resolved if they
were based on external parameters, which was the case in these tests.

To fix this, I've moved the downsampling feature  out of the
constructor and am calling it at a later stage, after all the
external dependencies are resolved.