Fixes map access in checkGraphState
If Xyce is built with the graph debugging flag enabled, there is a
map access call in checkGraphState that is incorrect.  Previously the
index was an integer, which is acceptable for the bracket operator.
However, the integer was changed to a pair to allow for the graph
to distinguish entries that are nodes versus devices.

https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/60