Handle case where infNorm() returns an invalid local index.
If the vector used in the infNorm() call to locate extrema is
empty, or all constant then infNorm() may return -1 for the
index.  If this index is used in name lookup, the name lookup
will fail.  The correct thing to do in this case it to not
do name lookup when infNorm() fails in extreama checking

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/574