Further fixes for local vs. global variation
This is related to this issue:

"Treat "local" and "global" variation for UQ methods in manner similar
to other simulators"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24

I thought that I fixed this issue last week.  However during  the QA
proecess several use cases were identified that didn't work correctly.

One problem was that the prior fix only affected device parameters
correctly and not other kinds of .params.

Another problem was that when AST nodes were replaced, the replacement
AST nodes were not always 100% resolved yet.

The 3rd problem was that in some cases, parameters that pre-replacement
were categorized as a constant parameter needed to be re-categorized as
a parameter that was variable.

All three of these issues are fixed in this commit.