Updated fix for local vs. global variation
"Treat "local" and "global" variation for UQ methods in manner similar
to other simulators"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24

This is an interim commit that will need to be cleaned up.  But, I think
I have this working now, including for parameters that are originally
declared as .param inside of subcircuits.  With the new methology, to
support local variaiton, .params (which are constant) are sometimes
converted to .global_params (which are variabl).

However, historically .global_params have not been allowed
inside of subcircuits.  Internally, .param and .global_param are much
more similar than they used to be.  But, some structural issues remain
that prevent .global_params from being 100% supported inside of
subcircuits.

One of those structural blockers was relevant to this issue.  When
peforming sampling, all the random parameters have to be updated at the
beginning of each sample step.  The specific values are generated in the
UQ functions, but the master list of parameters that must be updated is
maintained in the device manager.  .params are completely discarded once
parsing is done, and only .global_params that are actually used are
maintained in memory.

Prior to this commit, the code assumed that .global_params which
needed to be managed for sampling would come from the top level of the
netlist.  There was no mechanism for tracking .global_params that came
from subordinate subcircuits.

This commit adds the necessary code to do this, so it now works.  Doing
this required fully resolving the global  parameter name, in the same
manner that we resolve nodes and devices from subcircuits.