Checkpoint commit for issue 24
"Treat "local" and "global" variation for UQ methods in manner similar
to other simulators"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24

This is another interim commit.  The main updates are:

1.  "registerGlobalParams" now happens much earlier in pass 2, as it
is needed to handle models.  Models are set up on the fly as each
subcircuit is expanded, so Xyce can't wait to register subcircuit global
parameters with the device package until after subcircuits are done.

2.  All the processing of random parameters in subcircuits is
  now conditional.  Previously, it was done unconditionally.
The conditional in place now is "true" if any UQ commands are
present in the netlist and "false" otherwise.    Processing
random paraemters inside of subcircuits is more expensive,
so it should only be done if absolutely necessary.

3.  I've modified the code to ensure that random parameters inside of
  subcircuits, when converted to "global" parameters are doing it in
terms of the subcircuit instance, not the subcircuits definition.  This
means that the new, fully resolved name is based on the instance prefix.
It also means that for any subcircuit containing such a parameter, all
instances of that subcircuit must be re-resolved when expanded.
Unfortunatley, this is expensive.  I think this expense can be reduced,
but that will be the subject of future work.

4. This commit also modifies CircuitContext::resolvedParams,
to reduce the workload.

When expression-based parameters are resolved, much of the
work is associated with the "resolveStrings" function.  This function
identifies all the unresolved parameters that the original parameter
expression is dependent on, and then attempts to "resolve" them. (ie,
find them in the "resolvedParameters" or the "resolvedGlobalParameters"
containers.

If it fails to find any of them, then the function fails, and it is
(at some point) called again for the same parameter.  Eventually,
assuming the netlist doesn't have a mistake, one of the
"resolveParameter" function calls for that parameter will succeed.

In the previous implementation there was an inefficiency, in that it
threw away all the previous work.  So, if an expression depended on
(say) 10 external parameters, and the intial call successfully resolved
4 of them, and failed on the rest, then for the next call it should have
  only needed to resolve the remaining 6 params. Instead, it would do
the work to resolve all 10 again.    That wasn't the original intent of
this code, but it was what was happening.

This commit fixes this ineffiency.  For small netslist, this won't
matter much, but for very large netlists it should improve the setup
time.

5. This commit also corrects a visitor pattern mistake for funcOp
AST class.  The visitor pattern was, in some cases, not being
applied to function arguments.  This commit fixes that mistake.