Update a function for local vs global variation
"Treat "local" and "global" variation for UQ methods in manner similar
to other simulators"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24

Most of the functions in the circuit context that were updated for this
issue were various "resolve" functions, such as "resolveStrings",
"resolveParameter", etc.

One function that I overlooked, that also needed to be modified was
resolveParameterThatIsAdotFunc, which is a special case of
resolveParameter.  In practice, I think this change was probably
unncessary, but it is good to be consistent.  Also, I was able to make
this function slightly more efficient, just like I did with the original
resolveParameter.

Previously, both funnctions would re-allocate and re-parse expressions
every single time they were called, even if it was for a parameter that
already had a valid expression set up.   I fixed "resolveParameter" to
stop doing that, and in this commit I have now also fixed
resolveParameterThatIsAdotFunc in the same way.