Add method to generate expression string from AST
"Treat "local" and "global" variation for UQ methods in manner similar
to other simulators"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24

For some use cases, it is useful to be able to generate a new, parsable
expression string from an existing AST.  This commit adds this feature
to the expression library.  It also uses it as part of the
"newExpression::replaceParameterNode" function, which was created to
support local variation.  Previously, it updated
the expression string via substring search/replace, but that was very
error-prone.

This commmit also fixes the "processGlobalParam" function in the same
manner that I previously updated processParam and
processParamThatIsAFunc functions.  If the pararmeter being resolved
already has an allocated/parsed string, don't allocate a new one, just
use the old one.  This has a noticable impact on parser performance,
when there are large numbers of parameters.