Refactor parameter handling in device package
"Treat "local" and "global" variation for UQ methods in manner similar
to other simulators"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24

This commit gets rid of the device manager function,
DeviceMgr::updateDependentParameters_, and replaces it with several
smaller, more precise functions.

The reason updateDependentParameters_ needed to be replaced is that it
was too expensive in circuits with large numbers of parameters, and most
of the work that it was doing was unnecessary.

Specifically, it was updating nearly all global parameters every
time it was called, and it was being called from the "updateState"
function, meaning it was called at every Newton step.  This was bad
because:

1.  Some parameters were only "dependent" in the sense that
    they were dependent on other global parameters, and global
    parameters never change between Newton steps.    Xyce doesn't allow
    them to depend on solution variables.

2.  Some parameters were "dependent" in that they dependent on time or
    frequency.  But, similar to (1) above, time and frequency remain
    fixed during the Newton loop.  They only change between time steps, or
    AC/NOISE sweep steps.

3.  Also, global parameters (that were being updated in this function)
    only matter insomuch as they affect the values of device parameters.
    And, device parameters already incorporate change to global parameters
    in their update, since with the new expression library they share the
    same AST.  So, updating global param objects separately doesn't
    benefit them.

In fact, the main reason for global parameters to get a separate update
from device params is if the user has requested that they be output
on the .PRINT line.  And, not only that, but has requested them as a
raw parameter rather than an expression.  In other words, if the user
has ".PRINT TRAN param" on the .PRINT line, then the outputters will use
a "global parameter Op" class, which pulls the value from the device
package.  In that case, a correct value is needed, which is produced
during a global parameter udpate.

But, alternatively, if the user requests that same parameter inside of
an expression, (".PRINT TRAN {param}") then this isn't the case.  In
this instance, "{param}" will be processed by the expression library,
and a relevant AST is set up, and the outputter will instead use an
"expression Op" class.  In this case, it doesn't matter if the device
manager stores a value for {param} or not.

To address a lot of these issues, I now only update solution-dependent
params during Newton steps.  Time-depdendent params are updated at the
end of time steps, frequency-depedent params are updated at AC steps.
And, global params *if* they udpating, are updated during .STEP or
.SAMPLING iterations.  Most of these updates are done via the
notify/listener pattern.

This commit also includes a new boolean command line argument,
which can be used to toggle the new local variation feature.
By default it is "true", meaning local variation is enabled.

This commit includes a fair amount of commented out stuff, that I plan
to remove in a subsequent "cleanup" commit.