Refactor handling of subcircuit global params
This commit is related to:

"Treat "local" and "global" variation for UQ methods in manner similar
to other simulators"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24

My work on this issue so far has mostly focused on making this new
capability in serial, while not breaking any old capability in parallel.

However, it was always going to be necessary to add some
parallel-specific code eventually, for handling global parameters that
come from subcircuits.

Prior to my work on this issue, global parameters were not allowed
inside of subcircuits.  They were only allowed in the top-level netlist,
thus being referred to as "global" params.  However, the term "global
param" always really meant "variable params". ie, parameters allowed to
change.

As long as variable params were only allowed in the top level scope,
they were processed on the first netlist parser pass, and this pass was
entirely on proc 0, followed by a broadcast to the other processors.

By allowing local variation, this naturally leads to internal
subcircuit .params (which would normally be constants) to be elevated to
"global" or variable.  Handling this in serial was (ultimately)
straightforward, but it meant that new batches of global parameters were
being processed on the second parser pass.  And, due to the structure of
Xyce's second parser pass, these new parameters were not necessarily on
proc 0.  They were on whatever processor owned the particular
subcircuit they came from.

The order of operations has to be a little different on the second pass,
as in Xyce's parser design models are allocated on the fly, as needed,
and immediately set up in the device package.  To correctly be setup in
the device package, all variable params they might need must be
available.    So, subcircuit globals must be passed over a little
earlier in the process of the 2nd pass.

The UQ methods in Xyce currently assume that *all* uncertain params are
known on all processors.  Or, at least, all of them are known on proc 0.
This is because the updated random values for all uncertain parameters
are setup in one fell swoop in the UQ method functions.  For certain UQ
methods (such as quadrature based PCE, or sampling methods involving the
application of a coovariance matrix), they all have to be set up in the
same place.

So, this means that subcircuit "globals" must be tracked carefully and
then communicated to all processors.

This commit doesn't complete the parallel implementation, but it
re-organizes on subcircuit "globals" are handled, so that there is a
clear insertion point for this subcircuit global param communication.

This commit also includes a lot of cleanup.

This commit also includes some efficiency improvements in the parser.