Fix for local variation in parallel.  Final, for now.
"Treat "local" and "global" variation for UQ methods in manner similar
to other simulators"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24

The biggest change in this commit are the following:

1.  the subcircuitGlobalParams, which are created for local sampling,
    were not being set up in parallel, except on proc 0.  This is now
    fixed.

2.  Once they were handled on all processors, they were not consistent
    accross all processors.  Proc 0 had all of them, but proc != 0 only
    had subsets, that were encountered during parsing.  I added a
    function to broadcast them from proc 0, so all procs will have a
    full set.

For subcircuit globals that are subject to this broadcast, expression
-based global parameters could only be broadcast thru MPI as strings.
And then those strings are parsed by the expression library  to
create a new AST.

However, after this parsing is done, the expression isn't
necessarily resolved.   So, on processors other than proc 0, subcircuit
global params have to be re-resolved.  This feels a bit klunky, as
parameter resolutions are handled in the parser (specifically the
CircuitContext class).  But, it works. Fortunately, the resolution
process at this stage is (I think) cheaper, as all the dependent
.params were changed into their numerical values prior to generating
the expression strings to be broadcast.

So, in the strings that are used to re-generate expressions on the
receiving processor, the only "unresolved" params should be globals.
This should be a smaller list of things to resolve, and a smaller set of
params to search.