Fixes null char pointer being passed to string constructor
If FFTW returns a null char pointer that is passed to the
constructor to the string object, then an abort will be thrown before
exiting.  There should be a check that makes sure that this char
pointer is non-null before proceeding.

The commit adds the appropriate checks to avoid unnecessary aborts.

https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/40