More epression efficiency improvements
This is kind of related to this issue, but only kind of:

"Treat "local" and "global" variation for UQ methods in manner similar
to other simulators"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24

This commit includes some changes to the expression library and device
parameter handling that speed up behavioral circuits.  To some extent,
this commit is corrected an old mistake.

When expression are evaluated, usually the function that is called in
the AST is "dx2", which traverses the AST to evaluate the value and all
the needed derivatives for that expression.  This implementation
evaluates all the derivatives in one traversal, but doing this means
that some temporary arrays are used.    When I originally set this up,
these temporary arrays were of fixed size but this caused a bug for a
few use cases.  (see issue 369).   So, a couple of years ago I modified
it to allow these arrys to change size as needed.  Unfortunately, this
led to a lot of std::vector::size() and std::vector::resize() function
calls, which were expensive.

This commit handles the use case from issue 369 in a more efficient way,
minimizing the calls to size() and resize().  It is now roughly as fast
as it was before the fix to 369.

This commit contains a few other small efficiency improvements.
The expression group class for devices was checking if node names were
"Ground" equivalent every time the values were updated.
This wasn't necessary, as ground nodes get automagically converted
to numerical values (0.0) during parsing.