Add another std::sort to handling of random params
This is related to:

"Treat "local" and "global" variation for UQ methods in manner similar
to other simulators"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24

The function in the DeviceMgr that sets up the master list of random
parameters performs a bunch of std::sort calls to ensure that the list
of params is the same on each processor.   For various reasons,
the order isn't automatically the same on each proc.

However, I left out one that I probably should have included.  This
commit adds it.  This ensures also that the master list of parameters
will be in exactly the same in serial as it is in parallel as well.