Fix for seg fault on M1 mac
This fixes an issue that I caused with my work on this issue:

"Treat "local" and "global" variation for UQ methods in manner similar
to other simulators"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/24

As part of that work, I modified some of the parser code to be more
efficient.  For one particular change, I modified the following
CircuitContext functions:  resolveParameter, resolveGlobalParameter and
resolveParameterThatIsAFunc.

In all 3, I updated them so that they wouldn't allocate new expression
objects if there was already an expression object in the parameter being
resolved.  When they allocated brand new expressions every time, any
previous resolution work was thrown away, so this was a bad thing to do.

Unfortunately, I made a small mistake in the resolveGlobalParameter
function, which was causing a pointer to become invalid.  It didn't
break any of our normal testing platforms, but it did break the M1 mac.
This commit fixes the mistake.