Implement devConMap in BSIM CMG devices
Like a lot of simulators, Xyce performs a graph diagnostic to
detect if a circuit contains any nodes with "no DC path to ground".
Such nodes can sometimes cause convergence problems.

The code that performs this diagnostic relies on data structures
from each device model, which define what external nodes for that
device are connected to each other internally.  Most will be connected
but branches that are purely capacitive are not.  The data
structure is called "devConMap".

Unfortunately, there is a design flaw in Xyce, where if a device
does NOT implement this data structure, it will often produce
spurious warnings.  There is no enforcement, currently to ensure
that all devices have set this up.

Recently an user reported that a FinFET circuit, which used the
level=107 BSIM CMG model, was producing lot of incorrect warnings.
The reason was that the level=107 didn't have a devConMap.  As
it turned out, none of the CMG devices had it.

Like most modern models, the BSIM CMG models were generated from
a model compiler (ADMS), and that model compiler back-end currently
doesn't produce the devConMap code.  So, the code I am adding
in this commit was added manually.  If these devices are re-generated
via ADMS, or some other model compiler, it may disappear and have
to be re-added.

"The devConMap is not implemented in any of the BSIM CMG devices"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/611