Fix for segfault
This is for this bug:
https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/61
"Circuit causes Segfault"

This segfault was caused by a behavioral model that refers to a current
thru another "current source" Bsrc.  So in other words, something like
this:

This is valid, and runs fine:
B1 n1 n2 V={5}
B2 n3 n4 I={4*I(B1)}

This is not valid, and prior to this commit produces a segfault. It
*should* produce a clean error message:
B1 n1 n2 I={5}
B2 n3 n4 I={4*I(B1)}

Expressions in Bsrcs can depend on solution variables but not lead
currents.  This gets confusing with Bsrc's because they can either be
current sources or voltage sources.  If they are a voltage source (V=),
then they have an internal solution variable which is a current.

If they are a current source (I=), then they do NOT have an internal
variable.

The expression library attempts to trap for this issue.  So, it
understands I(V) is a valid current to reference (as it is a
solution variable) , and that I(R) is not a valid current to reference,
as it is a "lead current".

But, Bsrc's can be either, so it isn't possible to tell during parsing
if it is a valid current or not.  It has to be trapped later, during
external resolution of the Bsrc, when the GIDs/LIDS are looked up.
Xyce was not trapping this error correctly.  This commit adds an
appropriate error trap, and the seg fault will not happen.

The above invalid code will now exit with a reasonabl error message.