Fix for handling multiple .OPTIONS statements
This is related to this issue:

"Modify Xyce parser so that it can handle multiple .options statements
for the same component (for example, support .options timeint
statements)"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/571

Historically, we chose to have Xyce support ".OPTIONS <keyword>"
statements, instead of what SPICE does, with ".option".  ie, Xyce's
version is plural, and has a keyword indicating the part of the code
involved.  Hence, Xyce netlists have things like ".options timeint",
".options device" etc.

It was also designed assuming that all the options for a given keyword
would be in a single statement.  ie, that two instances of ".options
timeint" would not be allowed, or a typical use case.

In practice, however, most SPICE-style simulators allow the user to
specify as many ".OPTION" statements as they like.  So, Xyce users tend
to have this expectation.

Also, Xyce in recent years has had underfined behavior with respect to
this issue.  In recent builds of Xyce, if the user specifies multiple
".options device" lines, this is fine and Xyce uses all of them.
However, if the user specifies multiple ".options timeint" lines, Xyce
just uses the last one it finds during parsing, and throws away the
rest.  Neither behavior was planned, and it wasn't documented.  So, that
was bad.

With this commit, multiple .OPTIONS lines for the same keyword are
officially supported.  This means that somehting like this:

.options timeint reltol=1.0e-4
.options timeint abstol=1.0e-6

is the same as:

.options timeint reltol=1.0e-4 abstol=1.0e-6

Also, if the user specifies the same parameter more than once, Xyce will
issue a warning, and will use the first one it finds.  But, it will tell
you it is doing this.  So, if the user sets:

.options timeint reltol=1.0e-4
.options timeint reltol=1.0e-6

Xyce will issue a warning, and announce that it is using the first one,
reltol=1.0e-4.  The same will happen if the same option parameter is set
within the same line.  So, for example:

.options device voltlim=0  voltlim=1

will result in Xyce issuing a warning, and using the first value, 0.

Finally, while the section of Xyce that handles .OPTIONS statements also
handles a lot of other "dot" statements, such as .TRAN, only .OPTIONS
statements are allowed to be specified multiple times.  Multiple .TRAN
statements will not work.

And, among .OPTIONS statements, I had to make 2 exceptions.  .OPTIONS
OUTPUT and .OPTIONS RESTART both use a unique parameter,
"INITIAL_INTERVAL", which is specified as a space-separated list of
numbers, that ultimately get converted into time,value pairs.  However,
they get converted in such a way that all the times have the parameter
name "TIME" and all the values have the parameter name "INTERVAL".  So,
if the user specifies a list that is larger than one pair, then it has
multiple parmaeters of the same name, for these option blocks.  And,
all the identically named parameters in this case need to be preserved,
as they will all be used.

The easiest way to deal with this for now was to exclude the OUTPUT and
RESTART option blocks from this new "multi option" feature.