Add an #include to fix compilation failures
This is related to my work on:

"Modify Xyce parser so that it can handle multiple .options statements
for the same component (for example, support .options timeint
statements)"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/571

I did all that work using clang on my Mac.

However, it turns out that my changes caused build failures using gcc.
The issue is that an STL container was added to the IO::PkgOptionsMgr,
and it relied on a foward declaration (of the Util::OptionBlock class)
from N_UTL_fwd.h.  Apparently, the forward declaration was enough
information for clang to work, but it was not sufficient for gcc.

This commit fixes the issue by explicitly including N_UTL_OptionsBlock.h
in the package options manager.  This forces a lot of Xyce to be recompiled,
but it probably makes a certain kind of sense, given that the
IO::PkgOptionsMgr primarily manages containers of Util::OptionsBlock objects.