Improve source stepping for voltage source
The source stepping only works for DC sourcers. This makes it less
helpful to find a DC op, especially when there are transient sources
that have nonzero values at time 0. This commit improves the source
stepping for voltage sources for the case when both DC and transient
source types are specified for the voltage source. It improves the
convergence of source stepping for solving the DC op.

The source stepping only scales the default parameter in a source
which is DCV0. However, when both DC and transient source types are specified
for a source, none of the them were scaled during source stepping.
In this case, Xyce created both a DC and a transient source data. The
DC source data got scaled, but it is not used during source stepping.
This commit fixes the issue for the voltage source. The source stepping
now applies to the DC value when both DC and transient source types
are specified. It is shown to improve the DC convergence using source
stepping for some circuit

Xyce/backlog/project-backlog#583