Fix the source stepping for transient only source
The previous commit improves source stepping for the case when both
DC and transient source values are specified for a source. In this case,
the DC value is scaled during source stepping. However, when only transient
value is specified, this can lead to incorrect behavior. Xyce always
creats a DC source data, even when no DC value is specified. In this
case, Xyce scales the default 0 value DC source. If the transient source
has a nonzero value at time 0, this leads to unexpected behavior. This
commit fixes the issue. When a source only specifies transient value,
the transient source data is used during DC solve although it is not
scaled during source stepping. If homotopy is not used during DC
solve, the behavior is always correct

Xyce/backlog/project-backlog#583