Fix source stepping for instance class in voltage source
The commit applies the same fix for device master in voltage source
to instance class. The Vsrc has master class implemented. So the device
master is normally used in the device loading. The same fix is done
for device instance to make them consistent. It is used for certain
case, such as linear/nonlinear loading in circuit loader

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/583