Fix the device loading in current source
The current source implements a master class, but it is not
used during device loading. Instead, loading functions in the current
source instance class are used. This is confusing and inconsistent to
many other devices which implement device specific master class. It
is because the device specific master is not created in the factory of
current source. It is fixed now

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/583