chmod +x lib directory when making a release tarball
When copying over some of the Intel redistributable libraries,
we copy libraries with permissions 664. When we pack those
into the RPM, the RPM package manager fails to recognize them
as suitable for satisfying dependencies on them. We need to
ensure all libraries in the RPM are executable so that the
RPM package manager correctly picks them up when resolving
dependencies.