Fix for local variation via subckt arguments
This is related to this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/631
"Expression based local variation doesn't always work with subcircuit
arguments"

While doing QA for local variation, a use case was discovered that
didn't work correctly.  That use case was one in which the local random
operator (after processing) was applied to a subcircuit argument on the
"X" line.

Subcircuit "X" line arguments are handled separately from .params,
even though they ultimately (after processing) are stored in the same
parameter containers.  This is because they sit in a grey area between
two contexts.  Their dependencies can only depend on the calling
context, while their use is entirely within the subordinate subcircuit
context.  So, anyway, that is why they are treated a little differently,
and so this use case was not accounted for in the recent local
variation work.

This commit fixes that problem.  Now this code fragment will work
correctly:

.param main_var=aunif(5,5)
x1 2 0 sub1 arg1={main_var}

.subckt sub1 in out arg1=1
r1 in out {arg1}
.ends