Set up windows build to support shared library building
Modify the CMake build system to export all symbols by
default when building a shared-library version of Xyce
under Windows.  By default Windows will not export anything
in a shared library for external dynamic linking unless the
source code is modified to say which functions, classes and
data are ok to export.  For now Xyce's Windows builds will
do what the Linux and Mac builds do and just export all of
Xyce's symbols in the shared library version of Xyce

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/637