Install xyce-share.dll in bin dir
For a general "cmake --install ." and not a GUI installer
type of install, we were putting the Xyce dynamic library in
the lib directory where it is expected under MacOS and Linux.
Under Windows, the shared library needs to be in the bin
directory.  This is fixed by adding some conditional logic
to the CMakeLists.txt file

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/637