Add ctest based test to check Xyce after build
To test if a Xyce executable is successfully made and can link
to a dynamic library after it is built, I made a ctest based test
that is part of the build system.  This test doesn't need to use
GTest but leverages some of the cmake infrastructure added to
support GTest.  In gettig this test to work I made some other
tests that don't use GTest also work with dynamic libraries
under Windows.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/637